+++ sdk/notificationhubs/notification-hubs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.0.0-beta.2 (2022-08-17)
 
 - Removed internal dependencies of `@azure/core-amqp` and inlined code
-- Added Telemetry using the `X-MS-AZSDK-Telemetry` header.
+- Added Telemetry using the `x-ms-azsdk-telemetry` header.
 - Added the `User-Agent` for Notification Hubs `azsdk-js-messaging-notificationhubs`
 - Renamed the `/dist/index.js` to `/dis/index.cjs` for better CommonJS support.